Added feature "AutoHeight"
Added an autoHeight feature which can be enabled by setting the "autoHeight" option to true. If enabled the table will always grow (or shrink) so that all rows can be displayed
at once and no space is left in the table container.

This option needs a complete canvas resize in many situations and all rows are rendered at any time. This makes this option not suitable for large amounts of data.

Corresponding issue:

https://github.com/globexdesigns/doby-grid/issues/95