Fix sub rows from overwriting the model cache of the parent
The modified line appears to be using the parent ID to recursively write each sub row into the model cache, overwriting the parent and possibly some of the sub rows. This in itself doesn't cause any issues, but if you try to use remove() on the parent ID after expanding a child row, instead of removing the parent (and all children) from the table, it removes the last item in the table, because the get() call in remove() will always return -1 from its indexOf() check - thus, causing a negative slice from the items array. Using the same reference as the indexById return the line below resolves the problem.